+++ source/UberRides/RideRequestViewRequestingBehavior.swift
@@ -24,7 +24,11 @@
 
 import UberCore
 
-/// A RideRequesting object for requesting a ride via the RideRequestViewController
+/**
+ A RideRequesting object for requesting a ride via the RideRequestViewController
+ - Warning: The Ride Request Widget is deprecated, and will no longer work for new apps.
+ Existing apps have until 05/31/2018 to migrate. See the Uber API Changelog for more details.
+*/
 @objc(UBSDKRideRequestViewRequestingBehavior) public class RideRequestViewRequestingBehavior : NSObject {
     
     /// The UIViewController to present the RideRequestViewController over